Add time limit on order removal

+++ src/main/java/ch/wisv/events/core/tasks/OrderTaskScheduler.java
@@ -24,6 +24,9 @@ public class OrderTaskScheduler {
     /** Clean up task interval in seconds (30 minutes). */
     private static final int CLEAN_UP_TASK_INTERVAL_SECONDS = 1800;
 
+    /** Clean up orders after 7 days. */
+    private static final int CLEAN_UP_INTERVAL = 7;
+
     /** Max number of days a reservation is valid. */
     private static final int MAX_RESERVATION_DAYS = 3;
 
@@ -72,10 +75,12 @@ public class OrderTaskScheduler {
     @Scheduled(fixedRate = CLEAN_UP_TASK_INTERVAL_SECONDS * MILLISEC_IN_SEC)
     public void cleanUpTask() {
         orderService.getAllOrders().forEach(order -> {
-            OrderStatus[] cleanUpStatus = new OrderStatus[]{OrderStatus.ANONYMOUS, OrderStatus.ASSIGNED, OrderStatus.PENDING};
+            if (order.getCreatedAt().isBefore(LocalDateTime.now().minusDays(CLEAN_UP_INTERVAL))) {
+                OrderStatus[] cleanUpStatus = new OrderStatus[]{OrderStatus.ANONYMOUS, OrderStatus.ASSIGNED, OrderStatus.PENDING};
 
-            if (ArrayUtils.contains(cleanUpStatus, order.getStatus())) {
-                orderService.delete(order);
+                if (ArrayUtils.contains(cleanUpStatus, order.getStatus())) {
+                    orderService.delete(order);
+                }
             }
         });
     }
